sync

+++ app/src/main/java/com/example/geoapoyo/Models/ResponseServer/ServiciosExtra/R_ServiciosExtra_Coordenadar.java
@@ -1,6 +1,6 @@
 package com.example.geoapoyo.Models.ResponseServer.ServiciosExtra;
 
 public class R_ServiciosExtra_Coordenadar {
-    public int latitud;
-    public int longitud;
+    public double latitud;
+    public double longitud;
 }
